% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1999_2005.R
\docType{data}
\name{data_1999_2005}
\alias{data_1999_2005}
\title{Database of the sectors corresponding to the S&P 500 stocks}
\format{A matrix with first column stock symbols and second column stock sector}
\usage{
data(data_1999_2005)
}
\description{
Database of the sectors corresponding to the S&P 500 stocks in more detail.
}
\keyword{datasets}
