% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2006_2012.R
\docType{data}
\name{data_2006_2012}
\alias{data_2006_2012}
\title{Database of the net returns of the index S&P 500 and its underlying assets during 1/2006 - 12/2012}
\format{A list of two xts objects, namely, \code{X} and \code{SP500}, corresponding to the net returns of the assets and the index, respectively.}
\usage{
data(data_2006_2012)
}
\description{
Database of the net returns of the index S&P 500 and its underlying assets during the period 1/2006 - 12/2012.
Only the assets that were in the index during the whole period are included.
}
\keyword{datasets}
