% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2010_2015.R
\docType{data}
\name{data_2010_2015}
\alias{data_2010_2015}
\title{Database of the sectors corresponding to the S&P 500 stocks}
\format{A matrix with first column stock symbols and second column stock sector}
\usage{
data(data_2010_2015)
}
\description{
Database of the sectors corresponding to the S&P 500 stocks in more detail.
}
\keyword{datasets}
