% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDEX_2010.R
\docType{data}
\name{INDEX_2010}
\alias{INDEX_2010}
\title{Database of the net returns of the index S&P 500 and its underlying assets during the year 2010}
\format{A list of two \code{xts} objects, namely, \code{X} and \code{SP500}, corresponding to the net returns of the assets and the index, respectively.}
\usage{
data(INDEX_2010)
}
\description{
Database of the net returns of the index S&P 500 and its underlying assets during the year 2010.
Only the assets that were in the index during the whole period are included.
}
\keyword{datasets}
