% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaka1.R
\docType{data}
\name{malaka1}
\alias{malaka1}
\title{Database of the sectors corresponding to the S&P 500 stocks}
\format{A matrix with first column stock symbols and second column stock sector}
\usage{
data(malaka1)
}
\description{
Database of the sectors corresponding to the S&P 500 stocks in more detail.
}
\keyword{datasets}
