% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_sector_database.R
\docType{data}
\name{stock_sector_database}
\alias{stock_sector_database}
\title{Database of the sectors corresponding to the S&P 500 stocks}
\format{A matrix with first column stock symbols and second column stock sector
\describe{
  \item{symbol}{stock symbol}
  \item{sector}{sector of the stocks}
}}
\usage{
data(stock_sector_database)
}
\description{
Database of the sectors corresponding to the S&P 500 stocks
}
\examples{
data(stock_sector_database)
}
\keyword{datasets}
