% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2010_2015.R
\docType{data}
\name{data_2010_2015}
\alias{data_2010_2015}
\title{Database of the net returns of the index S&P 500 and its underlying assets}
\format{A list of two xts objects, namely 'X' and 'SP500', corresponding to the net returns of the assets and the index, respectively.}
\usage{
data(data_2010_2015)
}
\description{
Database of the net returns of the index S&P 500 and its underlying assets during the period 1/2010 - 12/2015.
Only the assets that where in the index during the whole period are included.
}
\keyword{datasets}
