% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndexTracking.R
\docType{package}
\name{sparseIndexTracking-package}
\alias{sparseIndexTracking-package}
\title{sparseIndexTracking: Design of Portfolio of Stocks to Tracks an Index}
\description{
[To be updated]
Computation of sparse eigenvectors of a matrix (aka sparse PCA)
with running time 2-3 orders of magnitude lower than existing methods and
better final performance in terms of recovery of sparsity pattern and
estimation of numerical values. Can handle covariance matrices as well as
data matrices with real or complex-valued entries. Different levels of
sparsity can be specified for each individual ordered eigenvector and the
method is robust in parameter selection. See vignette for a detailed
documentation and comparison, with several illustrative examples.
}
\section{Functions}{

\code{\link{xxx}}, \code{\link{xxx}}
}

\section{Help}{

For a quick help see the README \url{https://cran.r-project.org/web/packages/sparseIndexTracking/README.html}.

For more details see the vignette \url{https://cran.r-project.org/web/packages/sparseIndexTracking/vignettes/SparseEigenvectors.pdf}.
}

\references{
K. Benidis, Y. Feng, and D. P. Palomar, "Sparse Portfolios for High-Dimensional Financial Index Tracking,"
\emph{IEEE Transactions on Signal Processing}, vol. 66, no. 1, pp. 155-170, Jan. 2018.
}
\author{
Konstantinos Benidis and Daniel P. Palomar
}
